feat: add originated from field
This is a field to indicate where the sample comes from. It
has to be a collection object. If the sample origins from another
sample it should have the same originated from value.

Since the frontend isn't hooked up right atm I ended up with
just hardcode the value for now.