Fix/add analysis flow (#388)
* fix: map form to right values used by the backend

Still some values missing but this maps the values we
have.
Fields that the backend have that's not mapped:
- doneBy
- doneDate
- administrator
- completedBy
- completedDate
- reason

* fix: send null and not empty string when fields dosn't have value

When we're sending empty strings to the backend and the field is
option we're ending up with the `Some("")` values. This isn't
the behavior we want. This will send the null value if the string
is empty.

* fix: only show and send restriction if selected

The radio buttons controls if we should show and
send the restrictions form. By default it should
not be displayed to the user.

* fix: remove cancellation of restriction from add form

No point to add and cancel the restriction when adding
the analysis. This feature should be present in the
edit analysis!

* test: fix tests related to form changes

* fix: we don't have any hid value

* fix: remove history and updated/created fields

- We can't show any history for the analysis when
adding a new one.
- We don't have any registered by or updated by values
since the analusis hasn't been saved yet.

* style: code format on AnalysisAddComponent

* refactor: fix flow in AnalysisAddComponent

This should probably have been done better but it's good
enough for now! I guess we will see the right abstractions
when starting on the edit component.

* refactor: Update flow after changes to the add analysis component

* style: formatted the code

* fix: checked radio buttons + cleanup of default values shit

* fix: test missing field