feat: throw an error if lifecycle is used wrong
I thought that flow would have cached this kind of errors but it doesn't.
Since all the properties are optional it's perfectly ok to send in an
function since it doesn't break the contract. Since this have been changed
from taking an function to an object we' will not detect the faulty
usage when used wrong. This will at least give us a hard fault when
creating the component.