fix: analysis event page fails on 204 responses
We're failing to load the analysis event page when
the backend responds with a 204 (no content) response.
The reason we're failing it that we're expecting an
array from the backend but we're getting `undefined`.

This might be an issue other places we're just unpacking
the responses with the following syntax
`ajaxGet(url, token).map(({ response }) => (response));`

By doing so we're only considering the happy (200) path.