fix: set loading flag only on the request that's fired
This commit fixes two issues related to loading flags:

1) Correctly clear the store.
The origin solution cleared the store but had the assumption that
all the loading call should be fired. This isn't always the case but
it worked on initial load (or when we refreshed the page).

2) Set loading on only the actions that's fired.
Since we have a multiple action that set the loading state we need to
control them when they're triggered and before we actually call the api.
This is noe done before the api call by publishing the event to the
setLoading action. So now we should have a better control over what's
loading and not. This fixes an issue where some flags wasn't cleared
and resulted in a blank loading page.