fix: load data from root node when the store contains the root node
From the code it looks like we have tried to optimise the network calls
but not loading the root node twice. The issue is that we're not getting
any new props so we're not updating the state. This solution will use
the root node provided by the props if we're not loading it.