fix: clearing expiration date
The DatePicker component wasn't wired up correct. It did miss the onClear
props. That also introduced an issue where we want to blank out the value.

The DatePicker component is behaving a bit strange here and it looks like
it's not happy that we're clearing the value. Since we're wrapping the
component in the first place we probably should move the logic from the
restriction component into the date picker component.